+++ src/lib/index.ts
@@ -77,6 +77,7 @@ import { ViewComponent, ViewSwitcherComponent } from './view-switcher';
 import { IconComponent } from './icon';
 import { SelectDropdownComponent } from './select/select-dropdown';
 import { StackComponent } from './stack';
+import { ButtonAreaComponent } from './button-area';
 
 export * from './accordion';
 export * from './app-bar';
@@ -88,6 +89,7 @@ export * from './banner';
 export * from './bottom-sheet';
 export * from './busy-indicator';
 export * from './button';
+export * from './button-area';
 export * from './button-toggle';
 export * from './calendar';
 export * from './card';
@@ -153,6 +155,7 @@ const CUSTOM_ELEMENTS = [
   BottomSheetComponent,
   BusyIndicatorComponent,
   ButtonComponent,
+  ButtonAreaComponent,
   ButtonToggleComponent,
   ButtonToggleGroupComponent,
   CalendarComponent,